+++ autobahn/websocket/protocol.py
@@ -54,6 +54,7 @@ from autobahn.websocket.utf8validator import Utf8Validator
 from autobahn.websocket.xormasker import XorMaskerNull, create_xor_masker
 from autobahn.websocket.compress import PERMESSAGE_COMPRESSION_EXTENSION
 from autobahn.websocket.util import parse_url
+from autobahn.util import _maybe_tls_reason
 
 from six.moves import urllib
 import txaio
@@ -1092,7 +1093,12 @@ class WebSocketProtocol(object):
         else:
             if not self.wasClean:
                 if not self.droppedByMe and self.wasNotCleanReason is None:
-                    self.wasNotCleanReason = u'peer dropped the TCP connection without previous WebSocket closing handshake'
+                    reason_value = getattr(reason, 'value', None)
+                    reason_string = None if not reason_value else _maybe_tls_reason(reason_value)
+                    if reason_string:
+                        self.wasNotCleanReason = reason_string
+                    else:
+                        self.wasNotCleanReason = u'peer dropped the TCP connection without previous WebSocket closing handshake'
                 self._onClose(self.wasClean, WebSocketProtocol.CLOSE_STATUS_CODE_ABNORMAL_CLOSE, "connection was closed uncleanly (%s)" % self.wasNotCleanReason)
             else:
                 self._onClose(self.wasClean, self.remoteCloseCode, self.remoteCloseReason)
@@ -3673,7 +3679,6 @@ class WebSocketClientProtocol(WebSocketProtocol):
                         self._perMessageCompress = PMCE['PMCE'].create_from_response_accept(self.factory.isServer, accept)
 
                         self.websocket_extensions_in_use.append(self._perMessageCompress)
-                        print("PER MESSAGE COMPRESS {}".format(self._perMessageCompress))
 
                     else:
                         return self.failHandshake("server wants to use extension '%s' we did not request, haven't implemented or did not enable" % extension)