+++ autobahn/websocket/test/test_websocket_url.py
@@ -26,7 +26,7 @@
 
 from __future__ import absolute_import
 
-import unittest2 as unittest
+import unittest
 
 from autobahn.websocket.util import create_url, parse_url