+++ docker/x86_64/app/client_appsession_subscribe.py
@@ -1,0 +1,76 @@
+###############################################################################
+#
+# The MIT License (MIT)
+#
+# Copyright (c) Crossbar.io Technologies GmbH
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+#
+###############################################################################
+
+from __future__ import print_function
+from os import environ
+
+import os
+from twisted.internet import reactor
+from twisted.internet.defer import inlineCallbacks
+
+from autobahn.twisted.wamp import ApplicationSession, ApplicationRunner
+
+
+class ClientSession(ApplicationSession):
+    """
+    An application component that subscribes and receives events, and
+    stop after having received 5 events.
+    """
+
+    @inlineCallbacks
+    def onJoin(self, details):
+        print("session attached")
+        self.received = 0
+        sub = yield self.subscribe(self.on_event, u'com.myapp.hello')
+        print("Subscribed to com.myapp.hello with {}".format(sub.id))
+
+    def on_event(self, i):
+        print("Got event: {}".format(i))
+        self.received += 1
+        # self.config.extra for configuration, etc. (see [A])
+        if self.received > self.config.extra['max_events']:
+            print("Received enough events; disconnecting.")
+            self.leave()
+
+    def onDisconnect(self):
+        print("disconnected")
+        if reactor.running:
+            reactor.stop()
+
+
+if __name__ == '__main__':
+    import six
+    url = os.environ.get('CBURL', u'ws://localhost:8080/ws')
+    realm = os.environ.get('CBREALM', u'realm1')
+
+    # any extra info we want to forward to our ClientSession (in self.config.extra)
+    extra=dict(
+        max_events=5,  # [A] pass in additional configuration
+    )
+    runner = ApplicationRunner(url=url, realm=realm, extra=extra)
+    runner.run(ClientSession, auto_reconnect=True)
+
+